+++ game/items.py
@@ -85,11 +85,3 @@ class Flintlock(Item):
         self.skill = "guns"
         self.verb = "shoot"
         self.verb2 = "shoots"
-
-class GoldenClaymore(Item):
-    def __init__(self):
-        super().__init__("golden idol", 1000) #Note: price is in shillings (a silver coin, 20 per pound)
-        self.damage = (50,80)
-        self.skill = "swords"
-        self.verb = "slash"
-        self.verb2 = "slashes"
